Split Transforms into 2 more focused classes
